Update and/or remove dependencies, restructure app and improvements
