+++ Cargo.toml
@@ -7,9 +7,6 @@ authors = ["ChevyRay", "valflrt"]
 [dependencies]
 base64 = "0.13.0"
 byteorder = "1.4.3"
-eframe = "0.19.0"
-egui = "0.19.0"
-env_logger = "0.11.3"
-log = "0.4.21"
-rand = { version = "0.8.5", features = ["small_rng"] }
+eframe = "0.28.1"
+rand = {version = "0.8.5", features = ["small_rng"]}
 rayon = "1.9.0"
